Remove OrmMapper.java and TypedOrmMapper.java
